Extract sr_parser::training_data into separate files.
