+++ src/parser/CMakeLists.txt
@@ -7,6 +7,7 @@ add_subdirectory(tools)
 add_library(meta-parser sr_parser.cpp
                         state.cpp
                         state_analyzer.cpp
+                        training_data.cpp
                         transition.cpp
                         transition_map.cpp)
 target_link_libraries(meta-parser meta-parser-trees
